+++ cuid.d.ts
@@ -1,0 +1,5 @@
+declare module '@bugsnag/cuid' {
+  function cuid (): string
+
+  export default cuid
+}