+++ packages/platforms/react-native/android/src/main/java/com/bugsnag/android/performance/NativeBugsnagPerformanceImpl.java
@@ -1,0 +1,79 @@
+package com.bugsnag.android.performance;
+
+import android.content.Context;
+import android.content.pm.PackageInfo;
+import android.content.pm.PackageManager;
+import android.os.Build;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import com.bugsnag.android.performance.NativeBugsnagPerformanceSpec;
+import com.facebook.react.bridge.Arguments;
+import com.facebook.react.bridge.ReactApplicationContext;
+import com.facebook.react.bridge.WritableMap;
+
+public class NativeBugsnagPerformanceImpl extends NativeBugsnagPerformanceSpec {
+  
+  static final String NAME = "BugsnagReactNativePerformance";
+  
+  public NativeBugsnagPerformanceImpl(ReactApplicationContext reactContext) {
+    super(reactContext);
+  }
+
+  @NonNull
+  @Override
+  public String getName() {
+    return NAME;
+  }
+
+  @Override
+  public WritableMap getDeviceInfo() {
+    WritableMap map = Arguments.createMap();
+    try {
+      ReactApplicationContext reactContext = getReactApplicationContext();
+      String bundleIdentifier = reactContext.getPackageName();
+      map.putString("bundleIdentifier", bundleIdentifier);
+      PackageInfo packageInfo = reactContext.getPackageManager().getPackageInfo(bundleIdentifier, 0);
+      map.putString("versionCode", Integer.toString(packageInfo.versionCode));
+    } catch (Exception e) {
+      // ignore
+    }
+
+    String arch = null;
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+      String[] supportedAbis = Build.SUPPORTED_ABIS;
+      if (supportedAbis != null && supportedAbis.length > 0) {
+        arch = abiToArchitecture(supportedAbis[0]);
+      }
+    } else {
+      arch = abiToArchitecture(Build.CPU_ABI);
+    }
+
+    if (arch != null) {
+      map.putString("arch", arch);
+    }
+
+    map.putString("model", Build.MODEL);
+
+    return map;
+  }
+
+  @Nullable
+  private String abiToArchitecture(@Nullable String abi) {
+    if (abi == null) {
+      return null;
+    }
+
+    switch (abi) {
+      case "armeabi-v7a":
+        return "arm32";
+      case "arm64-v8a":
+        return "arm64";
+      case "x86":
+        return "x86";
+      case "x86_64":
+        return "amd64";
+      default:
+        return null;
+    }
+  }
+}