+++ packages/request-tracker/lib/network-request-callback.ts
@@ -1,0 +1,13 @@
+export interface NetworkRequestInfo {
+  url: string
+}
+
+export type NetworkRequestCallback <T extends NetworkRequestInfo> = (networkRequestInfo: T) => T | null
+
+export function defaultNetworkRequestCallback <T extends NetworkRequestInfo> (networkRequestInfo: T) {
+  return networkRequestInfo
+}
+
+export function isNetworkRequestCallback <T extends NetworkRequestInfo> (value: unknown): value is NetworkRequestCallback<T> {
+  return typeof value === 'function'
+}